[WIP] Basic working library
Used the amazing Soroush Falahati NvApiWrapper and Aitor González Fernández
nvapi.net codebases as a guide, along with Soroush's amazing script mentioned
here: https://stackoverflow.com/questions/13291783/how-to-get-the-id-memory-address-of-dll-function

This will let me get the memory addresses of NVIDIA functions as they are added in
future versions of the NVIDIA drivers. I do not need to create a full 1:1 NVAPI mapping
for all the functions listed, but instead I'll be only creating the minimal set of functions
that I need to support NVIDIA Surround and NVIDIA Mosaic. My main driver for doing
this work is to let me create updates and add features without needing to rely on a third
party to update the video driver libraries if I find issues.

I have to mention here that Soroush Falahati is absolutely incredible as a developer and
is a complete inspiration. I've reused quite a few bits of his NVIDIA NVAPI codebase within
this code, and I expect that I will use more (where allowed) as it is just so good! If you
haven't already, check out https://github.com/falahati/NvAPIWrapper as it's amazing!