Basic Working Mosiac config grab
Finally progress! I can now get the Mosiac display information from the NVIDIA driver and from that figure out if:
- mosiac mode is turned on right now
- If it was turned on, is there a topology that fits
- what the topology is
- what are the viewports within the mosaic topology

Next steps are to develop the setting ability; that is to take the a supplied configuration and then turn on (or off) the Mosaic topology and then use the WinLibrary to set the screen positions.