Fixed some bad windows CCD comparison logic
Removed the check for the Windows CCD being possible as
we cannot check for that. Mosaic changes the number of screens
which means that the Windows CCD checks will NEVER match.
Our only option is to assume that if the NVIDIA config is possible,
then the Windows CCD is possible too.