Disabled unused and non-working code
NvAPI_DISP_GetDisplayConfig doesn't work on the third pass, only passes one and two. This prevents us from getting the target parth information from the NVIDIA API. This doesn't matter to us as we actually already have the Windows CCD library that provides us that information. So we're using our library instead of the NVAPI to get that information.