Updated WinLibrary to the latest version
Updated WinLibrary to the latest version which added the GDI library funtionality to ensure that display refresh rates, BPC and display rotation were applied consistently within Windows. Some drivers appear to ignore the Windows CCD application of Display refresh rates, and required the addition of code to apply display settings via the GDI API as well in order for them to stick.