Finally fixed cloned profiles equality detection.
Was having issues when it would falsely NOT match when a clone profile was being applied in NVIDIA Surround mode. The profile would be applied, but wouldn't then subsequently match when it was checked against the stored profile. This now works, meaning that DisplayMagician will be able to tell when a cloned mode was successfully applied, and will reflect that withini the DisplayMagician GUI now.