Removed WinLibrary SetDisplayConfig validation
As we already get a valid, working display config from windows, it is unnecessary to validate it again. At the moment the validation is failing when a a slightly invalid config is tested, even though we provide a SDC_FORCE_MODE_ENUMERATION falg which will turn the config into a valid one. This effectively makes the validation useless, as a potentially valid config is incorrectly considered invalid, and fails. This now is not the case, and we only consider SetDisplayConfig failures when they actually happen (i.e. we try them and then track failure if they don't work). This is much more reliable.

We also now report each different log failure type to when using SetDisplayConfig.