Added a WinLibrary update config to refresh config
I *think* that NVIDIA surround changes are so great that WinLibrary gets confused and gives up. NVIDIA Surround affects the windows display layout so much. This change adds a windows config refresh after the NVIDIA config is applied to make sure that the winLibrary calls are using the latest updated active config that results from the NVIDIALibrary config changes.