First attempt at changing clone correction logic
The prior clone correction logic relied on HDR information being returned from the Windows CCD API. This wouldn't always work, leading to errors when trying to correct cloned profiles.

This change makes use of the fact that windows display configs are returned in path priority order. It grabs all physical display info, and finds the UID associated with them by Windows. It then goes through the returned display paths, and looks for the cloned display IDs that's invented by Windows. It then goes through the paths again, looking for the unmatched physical display UIDs. It then matches those two things up, in the order of how they were found, making a map from the faked clone ID to the real display ID.

You may ask why do all this work? Unfortuantely, for windows to load a cloned display, you need to set the cloned target display path entry ID to a specific high number in order for the cloned display to be recognised, and created. But the corresponding mode also needs to have it's ID changed, BUT TO THE ID OF THE REAL DISPLAY! This is such a pain in arse. And results in the code you see before you.

Typically, there is only one cloned screen, and so there is 1 matching 'missing' physical display UID. In some instances there may be multiple cloned displays, and this logic should now work for those instances!

Hopefully fixes #60 once and for all!