First working 1x1 setdisplaygrid based disable of mosaic
SUCCESS! I finally managed to get mosaic disabled using setdisplaygrid! We divide up the mosaic screens into individual 1x1 display grids, then apply it. This disables the mosaic, and the rest of the process continues.

We use NvAPI_Mosaic_EnableCurrentTopo to turn off mosaic if the setdisplaygrid method doesn't work. So we've even got a backup!