Disabled the taskbar setting and equality tests
WinLibrary will still record the taskbar position in display profiles, but will not actually try to set it, or use it in profile equality tests. This effectively will disable the taskbar feature for now. This was done as the feature still has lots of bugs and is not ready for release yet.