[WIP] Mostly working loading of DRS settings
Have completed the initial part of applying the DRS settings, but I've realised that none of the NVDRS_SETTING_V1 internal fields are stored. This means that the NVDRS_SETTING_V1 objects are all broken when they are loaded back from JSON. I will need to mark them as needed so they are stored properly.