Fixed broken equality testing when cloned display or after reboot
Turns out that Windows reorders some of the displays after cloning or while cloning or after a reboot. The equality testing wasn't taking care of that properly. It does now. I implemented equality test for lists which is order agnostic; it doesn't care what order the items are in the lists, it just cares that the objects are in both lists, and that there aren't any extra items in the list.

This change should fix #113 and #114.