Changed the GDI Settings to follow windows recommendations
Found that I'd been applying the GDI Display Settings in a way that isn't recommended by Microsoft. Found that we needed to call ChangeDisplaySettingsEx once for each display, but with the CDS_NORESET flag set so that it isn't immediately applied, and then once all settings have been staged, call ChangeDisplaySettingsEx once more set to all null. That last call triggers the actual modeset for all screens at once. This should avoid issues.

Additionally, I've added a test for each call to ChangeDisplaySettingsEx , and we skip setting that screen if it would cause an issue. This was done as 99% of the settings are in WindowsDisplayConfig anyway, so skipping applying GDI won't affect the overal user experience much. It's far better to aim for robustness in this case.

Should fix the latest error in #114